// Font-End/Font-end Admin-3/src/app/shared/models/B20AccDoc_Sales.ts
export class vB20AccDoc_Sales {

    id: number;
    stt: string;
    docStatus: number;
    docCode: string;
    docGroup: string;
    docDate: string;
    fiscalYear: string;
    customerCode: string;
    docNo: string;
    trackingNo: string;
    customerName: string;
    address: string;
    email: string;
    phone: string;
    note: string;
    currencyCode: string;
    exchangeRate: number;
    paymentMethod: string;
    paymentStatus: string;
    dueDate: number;
    dateDue: Date;

    totalOriginalAmount: number;
    totalAmount: number;
    totalOriginalAmount0: number;
    totalAmount0: number;
    totalOriginalAmount3: number;
    totalAmount3: number;
    totalOriginalAmount4: number;
    totalAmount4: number;
    isActive: boolean;
    createdBy: number;
    createdAt: Date;
    modifiedBy: number;
    modifiedAt: Date;
    bankCode: string;
    bankAccountNo: string;
    docStatusName: string;
    paymentName: string;
    payStatusName: string;
    taxCode: string;
    warehouseCode: string;
    taxRate: number;
    netWeight: number;
    location: string
    locationName: string;

    vB20AccDocSales_Json: Array<any>;

}